+++ empire.py
@@ -4,7 +4,11 @@ import sys
 
 from empire import arguments
 from empire.server.core.config import config_manager
-from empire.server.core.config.data_manager import sync_empire_compiler, sync_starkiller
+from empire.server.core.config.data_manager import (
+    sync_empire_compiler,
+    sync_plugin_registry,
+    sync_starkiller,
+)
 
 if __name__ == "__main__":
     args = arguments.args
@@ -16,5 +20,7 @@ if __name__ == "__main__":
     if args.subparser_name == "setup":
         sync_starkiller(config_manager.empire_config.starkiller)
         sync_empire_compiler(config_manager.empire_config.empire_compiler)
+        for registry in config_manager.empire_config.plugin_marketplace.registries:
+            sync_plugin_registry(registry)
 
     sys.exit(0)